imajidash-live/requirements.txt
asgiref==3.5.0
distlib==0.3.4
dj-database-url==0.5.0
Django==4.0.3
django-crispy-forms==1.14.0
django-heroku==0.3.1
django-phonenumber-field==6.1.0
filelock==3.6.0
gunicorn==20.1.0
phonenumbers==8.12.45
platformdirs==2.5.1
psycopg2==2.9.3
python-decouple==3.6
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
virtualenv==20.14.0
whitenoise==6.0.0
